Initial Upload
This is the initial Upload. Basically the goal for this point was to get all the content on the page and responsive. CSS styles will be added in the next phase, and the result will force the page to look much more like the pdf. This was just to get it all on the page and "working."