Phase Two: Design
At this point, the page is "done" in terms that everything is on the page, it has mobile functionality, and the color schemes are correct now. I am aware that the font is still not working, and will work in that in the next patch: Usability. That being said, at this point, any style changes should be done, with only minor changes due to functionality. All content is loaded and ready to go.