Final Phase
So at this point, the site is finished. MANY MANY bugs were squashed. Most notably the footer was redone to make it more responsive and friendly. Other changes throughout took place. Last note, I started this project as an .html without thinking about the fact that I'd have php on the site. The .php is what is actually live, but I uploaded a matching copy of the .html version so you can track changes made throughout. They are identical, but the .html is not being used anywhere.